+++ src/lib/index.ts
@@ -2,7 +2,9 @@ export { clickOutside } from '$lib/actions/click-outside.js'
 
 export {
   default as Connection,
-  type ConnectionStatus
+  type ConnectionState,
+  type ConnectionStatus,
+  connection
 } from '$lib/components/Connection.svelte'
 export { default as ToggleButton } from '$lib/components/ToggleButton.svelte'