+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased] - 
 
+### Added
+- Add global `Connection.connection` state to track connection status
+
 ### Fixed
 - Fix potential race condition with `.clearTimeout()` in `Connection` component